refactor: add repositories/ and services/
